requirements.txt
discord.py==1.7.3
PyMySQL==1.0.2
SQLAlchemy==1.4.35